% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta_squared.R
\name{petasq}
\alias{petasq}
\title{Partial Eta Squared}
\usage{
petasq(x, effect)
}
\arguments{
\item{x}{A call to \code{aov}, \code{ez::ezANOVA} or \code{afex::aov_ez} or
\code{afex::aov_car} or \code{afex::aov_4}}

\item{effect}{Character string indicating the name of the effect for which
the partial eta squared should be returned.}
}
\description{
Partial Eta Squared
}

