% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta_squared.R
\name{petasq_}
\alias{petasq_}
\title{Partial Eta Squared}
\usage{
petasq_(ss_effect, ss_error)
}
\arguments{
\item{ss_effect}{numeric, sum of squares of the effect}

\item{ss_error}{numeric, sum of squares of the corresponding error}
}
\description{
Calculate the partial eta squared effect size from sum of
squares.
\deqn{\eta_p^2 = \frac{SS_effect}{SS_effect + SS_error}}{partial eta squared
= SS_effect / (SS_effect + SS_error)}
}
